Renamed setting folder to settings
