+++ src/pages/settings/index.tsx
@@ -1,4 +1,4 @@
-import { SetStateAction, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { Formik } from "formik";
 import * as yup from "yup";
