+++ src/pages/settings/loadSettings.ts
@@ -16,9 +16,7 @@ const LoadSettings = (): null => {
       }
     })();
     // Clean up
-    return () => {
-      isMounted = false;
-    };
+    return () => { isMounted = false };
   }, []);
 
   return null;
